Interrupts now work, would still like to use meta-programming or a different way of specifying all of the interrupts than having them all typed out manually.
Templates/mixins aren't being used right now because it blows the executable size up from 50KiB -> 16MiB!
which is a little ridiculous.

The problems with the interrupts were:
 - Using wrong iret, it was outputting iretw instead of iretd
 - Forgetting to preserve registers while in the ISR

The interrupt hander should save all the stack registers, currently it does only a few.

Added a pin wheel to the clock/timer ISR to give some indication that everything is working